script/etl/requirements.txt
bs4==0.0.1
pandas==1.3.4
apache-airflow==2.3.2
python-dotenv==0.19.2
